requirements.txt
customtkinter
yt-dlp
python-vlc
mutagen
Pillow
keyboard
mouse
pyinstaller